Move logic from K8sdetail to ResizeCluster.tsx
