+++ packages/manager/src/features/Kubernetes/KubernetesClusterDetail/KubernetesClusterDetail.tsx
@@ -1,7 +1,6 @@
-import * as Bluebird from 'bluebird';
 import { getKubernetesClusterEndpoint } from 'linode-js-sdk/lib/kubernetes';
 import { APIError } from 'linode-js-sdk/lib/types';
-import { contains, equals, path, pathOr, remove, update } from 'ramda';
+import { path } from 'ramda';
 import * as React from 'react';
 import { RouteComponentProps, withRouter } from 'react-router-dom';
 import { compose } from 'recompose';
@@ -24,14 +23,10 @@ import KubeContainer, {
 import withTypes, { WithTypesProps } from 'src/containers/types.container';
 
 import { getAPIErrorOrDefault } from 'src/utilities/errorUtils';
-import scrollTo from 'src/utilities/scrollTo';
-import { getMonthlyPrice } from '.././kubeUtils';
-import { ExtendedCluster, PoolNodeWithPrice } from '.././types';
-import NodePoolPanel from '../CreateCluster/NodePoolPanel';
+import { ExtendedCluster } from '.././types';
 import KubeConfigPanel from './KubeConfigPanel';
 import KubernetesDialog from './KubernetesDialog';
 import KubeSummaryPanel from './KubeSummaryPanel';
-import NodePoolsDisplay from './NodePoolsDisplay';
 
 import Navigation from './DetailNavigation';
 
@@ -127,36 +122,9 @@ export const KubernetesClusterDetail: React.FunctionComponent<
     clustersLoading,
     lastUpdated,
     location,
-    nodePoolsLoading,
-    typesData,
-    typesError,
-    typesLoading,
-    requestClusterForStore,
-    requestKubernetesClusters,
-    requestNodePools,
-    updateCluster,
-    updateNodePool,
-    createNodePool,
-    deleteCluster,
-    deleteNodePool,
-    setKubernetesErrors,
-    ...routerProps
+    ...rest
   } = props;
 
-  const [editing, setEditing] = React.useState<boolean>(false);
-  /** Holds the local state of the cluster's node pools when editing */
-  const [pools, updatePools] = React.useState<PoolNodeWithPrice[]>([]);
-  /** When adding new pools in the NodePoolPanel component, use these variables. */
-  const [selectedType, setSelectedType] = React.useState<string | undefined>(
-    undefined
-  );
-  const [count, setCount] = React.useState<number>(1);
-  /** Form submission */
-  const [submitting, setSubmitting] = React.useState<boolean>(false);
-  const [generalError, setErrors] = React.useState<APIError[] | undefined>(
-    undefined
-  );
-  const [success, setSuccess] = React.useState<boolean>(false);
   /** Deletion confirmation modal */
   const [confirmationOpen, setConfirmation] = React.useState<boolean>(false);
   const [deleting, setDeleting] = React.useState<boolean>(false);
@@ -188,20 +156,6 @@ export const KubernetesClusterDetail: React.FunctionComponent<
         });
     }
 
-    /**
-     * If we're navigating from the action menu on the cluster list page,
-     * we want to start with editing mode active.
-     */
-    const isEditing = pathOr(
-      false,
-      ['history', 'location', 'state', 'editing'],
-      props
-    );
-
-    if (isEditing !== editing) {
-      toggleEditing();
-    }
-
     const interval = setInterval(
       () => props.requestNodePools(+props.match.params.clusterID),
       10000
@@ -222,8 +176,8 @@ export const KubernetesClusterDetail: React.FunctionComponent<
 
   if (
     (clustersLoading && lastUpdated === 0) ||
-    nodePoolsLoading ||
-    typesLoading
+    props.nodePoolsLoading ||
+    props.typesLoading
   ) {
     return <CircleProgress />;
   }
@@ -231,188 +185,6 @@ export const KubernetesClusterDetail: React.FunctionComponent<
   if (cluster === null) {
     return null;
   }
-
-  /**
-   * These three handlers update the local pools state in the event of an error. If an update
-   * is fully successful, we'll exit editing mode, the table will show
-   * the current Redux state of the cluster, and none of this will matter.
-   *
-   * If, however, some requests fail while others succeed, we want to show
-   * error messages for actions that failed while updating the local form
-   * state for actions that succeeded (so that e.g. a pending pool that has been added no longer has the
-   * "pending pool" styles).
-   */
-  const handleError = (pool: PoolNodeWithPrice, error: APIError[]) => {
-    const poolIdx = pools.findIndex(thisPool => thisPool.id === pool.id);
-    updatePool(poolIdx, { ...pool, _error: error });
-    return Promise.reject(error);
-  };
-
-  const handleAddSuccess = (pool: PoolNodeWithPrice) => {
-    const poolIdx = pools.findIndex(thisPool => thisPool.id === pool.id);
-    updatePool(poolIdx, {
-      ...pool,
-      queuedForAddition: false,
-      queuedForDeletion: false
-    });
-  };
-
-  const handleDeleteSuccess = (pool: PoolNodeWithPrice) => {
-    const poolIdx = pools.findIndex(thisPool => thisPool.id === pool.id);
-    if (poolIdx) {
-      updatePools(prevPools => {
-        return remove(poolIdx, 1, prevPools);
-      });
-    }
-  };
-
-  const submitForm = () => {
-    /** If the user hasn't made any input, there's nothing to submit. */
-    if (equals(pools, cluster.node_pools)) {
-      setEditing(false);
-      return;
-    }
-    /** Fasten your seat belts... */
-    setSubmitting(true);
-    setErrors(undefined);
-    setSuccess(false);
-    Bluebird.map(pools, thisPool => {
-      if (thisPool.queuedForAddition) {
-        // This pool doesn't exist and needs to be created through the API.
-        return props
-          .createNodePool({
-            clusterID: cluster.id,
-            count: thisPool.count,
-            type: thisPool.type
-          })
-          .then(() => handleAddSuccess(thisPool))
-          .catch(e => handleError(thisPool, e));
-      } else if (thisPool.queuedForDeletion) {
-        // Marked for deletion
-        return props
-          .deleteNodePool({
-            clusterID: cluster.id,
-            nodePoolID: thisPool.id
-          })
-          .then(() => handleDeleteSuccess(thisPool))
-          .catch(e => handleError(thisPool, e));
-      } else if (!contains(thisPool, cluster.node_pools)) {
-        /** @todo contains() is deprecated in the next version of Ramda (0.26+). Replace with includes() if we ever upgrade. */
-
-        // User has adjusted the count for this pool. Needs to be pushed through to the API.
-        return props
-          .updateNodePool({
-            clusterID: cluster.id,
-            nodePoolID: thisPool.id,
-            count: thisPool.count,
-            type: thisPool.type
-          })
-          .catch(e => handleError(thisPool, e));
-      } else {
-        // Nothing has changed about this node, so don't make any requests.
-        return;
-      }
-    })
-      .then(() => {
-        setSuccess(true);
-        setSubmitting(false);
-        setEditing(false);
-      })
-      .catch(err => {
-        setErrors(
-          getAPIErrorOrDefault(
-            err,
-            'Some actions could not be completed. Please try again.'
-          )
-        );
-        setSubmitting(false);
-      });
-  };
-
-  const updatePool = (poolIdx: number, updatedPool: PoolNodeWithPrice) => {
-    const updatedPoolWithPrice = {
-      ...updatedPool,
-      totalMonthlyPrice: getMonthlyPrice(
-        updatedPool.type,
-        updatedPool.count,
-        typesData || []
-      )
-    };
-    updatePools(prevPools => update(poolIdx, updatedPoolWithPrice, prevPools));
-  };
-
-  const resetFormState = () => {
-    updatePools(cluster.node_pools);
-  };
-
-  const toggleEditing = () => {
-    /**
-     * Regardless of whether we're going into or out of editing mode,
-     * reset everything.
-     */
-
-    updatePools(cluster.node_pools);
-    setEditing(!editing);
-    setSuccess(false);
-    setErrors(undefined);
-  };
-
-  const handleAddNodePool = (pool: PoolNodeWithPrice) => {
-    if (editing) {
-      /** We're already in editing mode, so the list of pool nodes should be accurate */
-      updatePools(prevPools => {
-        const newPools = [
-          ...prevPools,
-          {
-            ...pool,
-            queuedForAddition: true
-          }
-        ];
-        scrollTo();
-        return newPools;
-      });
-    } else {
-      /** From a static state, adding a node pool should trigger editing state */
-      scrollTo();
-      toggleEditing();
-      /** Make sure the list of node pools is correct */
-      updatePools([
-        ...cluster.node_pools,
-        {
-          ...pool,
-          queuedForAddition: true
-        }
-      ]);
-    }
-  };
-
-  const handleDeletePool = (poolIdx: number) => {
-    updatePools(prevPools => {
-      const poolToDelete = path<PoolNodeWithPrice>([poolIdx], prevPools);
-      if (poolToDelete) {
-        if (poolToDelete.queuedForAddition) {
-          /**
-           * This is a new pool in local state that doesn't exist as far as the API is concerned.
-           * It can be directly removed.
-           */
-          return remove(poolIdx, 1, prevPools);
-        } else {
-          /**
-           * This is a "real" node that we don't want users to accidentally delete. Mark it for deletion
-           * (it will be handled on form submission). If the user has already marked this for deletion
-           * and clicks on "Remove Delete", remove the queuedForDeletion tag.
-           */
-          const withMarker = {
-            ...poolToDelete,
-            queuedForDeletion: !Boolean(poolToDelete.queuedForDeletion)
-          };
-          return update(poolIdx, withMarker, prevPools);
-        }
-      } else {
-        return prevPools;
-      }
-    });
-  };
 
   const handleDeleteCluster = () => {
     setDeleting(true);
@@ -492,7 +264,13 @@ export const KubernetesClusterDetail: React.FunctionComponent<
           md={9}
           className={classes.sectionMain}
         >
-          <Navigation location={location} {...routerProps} />
+          <Navigation
+            location={location}
+            cluster={cluster}
+            updateCluster={props.updateCluster}
+            updateNodePool={props.updateNodePool}
+            {...rest}
+          />
           {/* <Grid item xs={12}>
             <NodePoolsDisplay
               submittingForm={submitting}
